deps/v8/test/webkit/fast/regex/unicodeCaseInsensitive-expected.txt

# Copyright 2013 the V8 project authors. All rights reserved.
# Copyright (C) 2005, 2006, 2007, 2008, 2009 Apple Inc. All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions
# are met:
# 1.  Redistributions of source code must retain the above copyright
#     notice, this list of conditions and the following disclaimer.
# 2.  Redistributions in binary form must reproduce the above copyright
#     notice, this list of conditions and the following disclaimer in the
#     documentation and/or other materials provided with the distribution.
#
# THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

https://bugs.webkit.org/show_bug.cgi?id=82063

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS /ΣΤΙΓΜΑΣ/i.test("στιγμας") is true
PASS /ΔΣΔ/i.test("δςδ") is true
PASS /ς/i.test("σ") is true
PASS /σ/i.test("ς") is true
PASS /἖/i.test("἖") is true
PASS /Ǆ/i.test("Ǆ") is true
PASS /[ǃ-ǅ]/i.test("Ǆ") is true
PASS /Ǆ/i.test("ǅ") is true
PASS /[ǃ-ǅ]/i.test("ǅ") is true
PASS /Ǆ/i.test("ǆ") is true
PASS /[ǃ-ǅ]/i.test("ǆ") is true
PASS /ǅ/i.test("Ǆ") is true
PASS /[Ǆ-ǆ]/i.test("Ǆ") is true
PASS /ǅ/i.test("ǅ") is true
PASS /[Ǆ-ǆ]/i.test("ǅ") is true
PASS /ǅ/i.test("ǆ") is true
PASS /[Ǆ-ǆ]/i.test("ǆ") is true
PASS /ǆ/i.test("Ǆ") is true
PASS /[ǅ-Ǉ]/i.test("Ǆ") is true
PASS /ǆ/i.test("ǅ") is true
PASS /[ǅ-Ǉ]/i.test("ǅ") is true
PASS /ǆ/i.test("ǆ") is true
PASS /[ǅ-Ǉ]/i.test("ǆ") is true
PASS /Ǉ/i.test("Ǉ") is true
PASS /[ǆ-ǈ]/i.test("Ǉ") is true
PASS /Ǉ/i.test("ǈ") is true
PASS /[ǆ-ǈ]/i.test("ǈ") is true
PASS /Ǉ/i.test("ǉ") is true
PASS /[ǆ-ǈ]/i.test("ǉ") is true
PASS /ǈ/i.test("Ǉ") is true
PASS /[Ǉ-ǉ]/i.test("Ǉ") is true
PASS /ǈ/i.test("ǈ") is true
PASS /[Ǉ-ǉ]/i.test("ǈ") is true
PASS /ǈ/i.test("ǉ") is true
PASS /[Ǉ-ǉ]/i.test("ǉ") is true
PASS /ǉ/i.test("Ǉ") is true
PASS /[ǈ-Ǌ]/i.test("Ǉ") is true
PASS /ǉ/i.test("ǈ") is true
PASS /[ǈ-Ǌ]/i.test("ǈ") is true
PASS /ǉ/i.test("ǉ") is true
PASS /[ǈ-Ǌ]/i.test("ǉ") is true
PASS /Ǌ/i.test("Ǌ") is true
PASS /[ǉ-ǋ]/i.test("Ǌ") is true
PASS /Ǌ/i.test("ǋ") is true
PASS /[ǉ-ǋ]/i.test("ǋ") is true
PASS /Ǌ/i.test("ǌ") is true
PASS /[ǉ-ǋ]/i.test("ǌ") is true
PASS /ǋ/i.test("Ǌ") is true
PASS /[Ǌ-ǌ]/i.test("Ǌ") is true
PASS /ǋ/i.test("ǋ") is true
PASS /[Ǌ-ǌ]/i.test("ǋ") is true
PASS /ǋ/i.test("ǌ") is true
PASS /[Ǌ-ǌ]/i.test("ǌ") is true
PASS /ǌ/i.test("Ǌ") is true
PASS /[ǋ-Ǎ]/i.test("Ǌ") is true
PASS /ǌ/i.test("ǋ") is true
PASS /[ǋ-Ǎ]/i.test("ǋ") is true
PASS /ǌ/i.test("ǌ") is true
PASS /[ǋ-Ǎ]/i.test("ǌ") is true
PASS /Ǳ/i.test("Ǳ") is true
PASS /[ǰ-ǲ]/i.test("Ǳ") is true
PASS /Ǳ/i.test("ǲ") is true
PASS /[ǰ-ǲ]/i.test("ǲ") is true
PASS /Ǳ/i.test("ǳ") is true
PASS /[ǰ-ǲ]/i.test("ǳ") is true
PASS /ǲ/i.test("Ǳ") is true
PASS /[Ǳ-ǳ]/i.test("Ǳ") is true
PASS /ǲ/i.test("ǲ") is true
PASS /[Ǳ-ǳ]/i.test("ǲ") is true
PASS /ǲ/i.test("ǳ") is true
PASS /[Ǳ-ǳ]/i.test("ǳ") is true
PASS /ǳ/i.test("Ǳ") is true
PASS /[ǲ-Ǵ]/i.test("Ǳ") is true
PASS /ǳ/i.test("ǲ") is true
PASS /[ǲ-Ǵ]/i.test("ǲ") is true
PASS /ǳ/i.test("ǳ") is true
PASS /[ǲ-Ǵ]/i.test("ǳ") is true
PASS /Β/i.test("Β") is true
PASS /[Α-Γ]/i.test("Β") is true
PASS /Β/i.test("β") is true
PASS /[Α-Γ]/i.test("β") is true
PASS /Β/i.test("ϐ") is true
PASS /[Α-Γ]/i.test("ϐ") is true
PASS /β/i.test("Β") is true
PASS /[α-γ]/i.test("Β") is true
PASS /β/i.test("β") is true
PASS /[α-γ]/i.test("β") is true
PASS /β/i.test("ϐ") is true
PASS /[α-γ]/i.test("ϐ") is true
PASS /ϐ/i.test("Β") is true
PASS /[Ϗ-ϑ]/i.test("Β") is true
PASS /ϐ/i.test("β") is true
PASS /[Ϗ-ϑ]/i.test("β") is true
PASS /ϐ/i.test("ϐ") is true
PASS /[Ϗ-ϑ]/i.test("ϐ") is true
PASS /Ε/i.test("Ε") is true
PASS /[Δ-Ζ]/i.test("Ε") is true
PASS /Ε/i.test("ε") is true
PASS /[Δ-Ζ]/i.test("ε") is true
PASS /Ε/i.test("ϵ") is true
PASS /[Δ-Ζ]/i.test("ϵ") is true
PASS /ε/i.test("Ε") is true
PASS /[δ-ζ]/i.test("Ε") is true
PASS /ε/i.test("ε") is true
PASS /[δ-ζ]/i.test("ε") is true
PASS /ε/i.test("ϵ") is true
PASS /[δ-ζ]/i.test("ϵ") is true
PASS /ϵ/i.test("Ε") is true
PASS /[ϴ-϶]/i.test("Ε") is true
PASS /ϵ/i.test("ε") is true
PASS /[ϴ-϶]/i.test("ε") is true
PASS /ϵ/i.test("ϵ") is true
PASS /[ϴ-϶]/i.test("ϵ") is true
PASS /Θ/i.test("Θ") is true
PASS /[Η-Ι]/i.test("Θ") is true
PASS /Θ/i.test("θ") is true
PASS /[Η-Ι]/i.test("θ") is true
PASS /Θ/i.test("ϑ") is true
PASS /[Η-Ι]/i.test("ϑ") is true
PASS /θ/i.test("Θ") is true
PASS /[η-ι]/i.test("Θ") is true
PASS /θ/i.test("θ") is true
PASS /[η-ι]/i.test("θ") is true
PASS /θ/i.test("ϑ") is true
PASS /[η-ι]/i.test("ϑ") is true
PASS /ϑ/i.test("Θ") is true
PASS /[ϐ-ϒ]/i.test("Θ") is true
PASS /ϑ/i.test("θ") is true
PASS /[ϐ-ϒ]/i.test("θ") is true
PASS /ϑ/i.test("ϑ") is true
PASS /[ϐ-ϒ]/i.test("ϑ") is true
PASS /ͅ/i.test("ͅ") is true
PASS /[̈́-͆]/i.test("ͅ") is true
PASS /ͅ/i.test("Ι") is true
PASS /[̈́-͆]/i.test("Ι") is true
PASS /ͅ/i.test("ι") is true
PASS /[̈́-͆]/i.test("ι") is true
PASS /ͅ/i.test("ι") is true
PASS /[̈́-͆]/i.test("ι") is true
PASS /Ι/i.test("ͅ") is true
PASS /[Θ-Κ]/i.test("ͅ") is true
PASS /Ι/i.test("Ι") is true
PASS /[Θ-Κ]/i.test("Ι") is true
PASS /Ι/i.test("ι") is true
PASS /[Θ-Κ]/i.test("ι") is true
PASS /Ι/i.test("ι") is true
PASS /[Θ-Κ]/i.test("ι") is true
PASS /ι/i.test("ͅ") is true
PASS /[θ-κ]/i.test("ͅ") is true
PASS /ι/i.test("Ι") is true
PASS /[θ-κ]/i.test("Ι") is true
PASS /ι/i.test("ι") is true
PASS /[θ-κ]/i.test("ι") is true
PASS /ι/i.test("ι") is true
PASS /[θ-κ]/i.test("ι") is true
PASS /ι/i.test("ͅ") is true
PASS /[᾽-᾿]/i.test("ͅ") is true
PASS /ι/i.test("Ι") is true
PASS /[᾽-᾿]/i.test("Ι") is true
PASS /ι/i.test("ι") is true
PASS /[᾽-᾿]/i.test("ι") is true
PASS /ι/i.test("ι") is true
PASS /[᾽-᾿]/i.test("ι") is true
PASS /Κ/i.test("Κ") is true
PASS /[Ι-Λ]/i.test("Κ") is true
PASS /Κ/i.test("κ") is true
PASS /[Ι-Λ]/i.test("κ") is true
PASS /Κ/i.test("ϰ") is true
PASS /[Ι-Λ]/i.test("ϰ") is true
PASS /κ/i.test("Κ") is true
PASS /[ι-λ]/i.test("Κ") is true
PASS /κ/i.test("κ") is true
PASS /[ι-λ]/i.test("κ") is true
PASS /κ/i.test("ϰ") is true
PASS /[ι-λ]/i.test("ϰ") is true
PASS /ϰ/i.test("Κ") is true
PASS /[ϯ-ϱ]/i.test("Κ") is true
PASS /ϰ/i.test("κ") is true
PASS /[ϯ-ϱ]/i.test("κ") is true
PASS /ϰ/i.test("ϰ") is true
PASS /[ϯ-ϱ]/i.test("ϰ") is true
PASS /µ/i.test("µ") is true
PASS /[´-¶]/i.test("µ") is true
PASS /µ/i.test("Μ") is true
PASS /[´-¶]/i.test("Μ") is true
PASS /µ/i.test("μ") is true
PASS /[´-¶]/i.test("μ") is true
PASS /Μ/i.test("µ") is true
PASS /[Λ-Ν]/i.test("µ") is true
PASS /Μ/i.test("Μ") is true
PASS /[Λ-Ν]/i.test("Μ") is true
PASS /Μ/i.test("μ") is true
PASS /[Λ-Ν]/i.test("μ") is true
PASS /μ/i.test("µ") is true
PASS /[λ-ν]/i.test("µ") is true
PASS /μ/i.test("Μ") is true
PASS /[λ-ν]/i.test("Μ") is true
PASS /μ/i.test("μ") is true
PASS /[λ-ν]/i.test("μ") is true
PASS /Π/i.test("Π") is true
PASS /[Ο-Ρ]/i.test("Π") is true
PASS /Π/i.test("π") is true
PASS /[Ο-Ρ]/i.test("π") is true
PASS /Π/i.test("ϖ") is true
PASS /[Ο-Ρ]/i.test("ϖ") is true
PASS /π/i.test("Π") is true
PASS /[ο-ρ]/i.test("Π") is true
PASS /π/i.test("π") is true
PASS /[ο-ρ]/i.test("π") is true
PASS /π/i.test("ϖ") is true
PASS /[ο-ρ]/i.test("ϖ") is true
PASS /ϖ/i.test("Π") is true
PASS /[ϕ-ϗ]/i.test("Π") is true
PASS /ϖ/i.test("π") is true
PASS /[ϕ-ϗ]/i.test("π") is true
PASS /ϖ/i.test("ϖ") is true
PASS /[ϕ-ϗ]/i.test("ϖ") is true
PASS /Ρ/i.test("Ρ") is true
PASS /[Π-΢]/i.test("Ρ") is true
PASS /Ρ/i.test("ρ") is true
PASS /[Π-΢]/i.test("ρ") is true
PASS /Ρ/i.test("ϱ") is true
PASS /[Π-΢]/i.test("ϱ") is true
PASS /ρ/i.test("Ρ") is true
PASS /[π-ς]/i.test("Ρ") is true
PASS /ρ/i.test("ρ") is true
PASS /[π-ς]/i.test("ρ") is true
PASS /ρ/i.test("ϱ") is true
PASS /[π-ς]/i.test("ϱ") is true
PASS /ϱ/i.test("Ρ") is true
PASS /[ϰ-ϲ]/i.test("Ρ") is true
PASS /ϱ/i.test("ρ") is true
PASS /[ϰ-ϲ]/i.test("ρ") is true
PASS /ϱ/i.test("ϱ") is true
PASS /[ϰ-ϲ]/i.test("ϱ") is true
PASS /Σ/i.test("Σ") is true
PASS /[΢-Τ]/i.test("Σ") is true
PASS /Σ/i.test("ς") is true
PASS /[΢-Τ]/i.test("ς") is true
PASS /Σ/i.test("σ") is true
PASS /[΢-Τ]/i.test("σ") is true
PASS /ς/i.test("Σ") is true
PASS /[ρ-σ]/i.test("Σ") is true
PASS /ς/i.test("ς") is true
PASS /[ρ-σ]/i.test("ς") is true
PASS /ς/i.test("σ") is true
PASS /[ρ-σ]/i.test("σ") is true
PASS /σ/i.test("Σ") is true
PASS /[ς-τ]/i.test("Σ") is true
PASS /σ/i.test("ς") is true
PASS /[ς-τ]/i.test("ς") is true
PASS /σ/i.test("σ") is true
PASS /[ς-τ]/i.test("σ") is true
PASS /Φ/i.test("Φ") is true
PASS /[Υ-Χ]/i.test("Φ") is true
PASS /Φ/i.test("φ") is true
PASS /[Υ-Χ]/i.test("φ") is true
PASS /Φ/i.test("ϕ") is true
PASS /[Υ-Χ]/i.test("ϕ") is true
PASS /φ/i.test("Φ") is true
PASS /[υ-χ]/i.test("Φ") is true
PASS /φ/i.test("φ") is true
PASS /[υ-χ]/i.test("φ") is true
PASS /φ/i.test("ϕ") is true
PASS /[υ-χ]/i.test("ϕ") is true
PASS /ϕ/i.test("Φ") is true
PASS /[ϔ-ϖ]/i.test("Φ") is true
PASS /ϕ/i.test("φ") is true
PASS /[ϔ-ϖ]/i.test("φ") is true
PASS /ϕ/i.test("ϕ") is true
PASS /[ϔ-ϖ]/i.test("ϕ") is true
PASS /Ṡ/i.test("Ṡ") is true
PASS /[ṟ-ṡ]/i.test("Ṡ") is true
PASS /Ṡ/i.test("ṡ") is true
PASS /[ṟ-ṡ]/i.test("ṡ") is true
PASS /Ṡ/i.test("ẛ") is true
PASS /[ṟ-ṡ]/i.test("ẛ") is true
PASS /ṡ/i.test("Ṡ") is true
PASS /[Ṡ-Ṣ]/i.test("Ṡ") is true
PASS /ṡ/i.test("ṡ") is true
PASS /[Ṡ-Ṣ]/i.test("ṡ") is true
PASS /ṡ/i.test("ẛ") is true
PASS /[Ṡ-Ṣ]/i.test("ẛ") is true
PASS /ẛ/i.test("Ṡ") is true
PASS /[ẚ-ẜ]/i.test("Ṡ") is true
PASS /ẛ/i.test("ṡ") is true
PASS /[ẚ-ẜ]/i.test("ṡ") is true
PASS /ẛ/i.test("ẛ") is true
PASS /[ẚ-ẜ]/i.test("ẛ") is true
PASS /Ϗ/i.test("Ϗ") is true
PASS /ϗ/i.test("Ϗ") is true
PASS /Ϗ/i.test("ϗ") is true
PASS /ϗ/i.test("ϗ") is true
PASS /ἑ/i.test("ἑ") is true
PASS /Ἑ/i.test("ἑ") is true
PASS /ἑ/i.test("Ἑ") is true
PASS /Ἑ/i.test("Ἑ") is true
PASS /҉/i.test("Ҋ") is false
PASS /Ҋ/i.test("Ҋ") is true
PASS /ҋ/i.test("Ҋ") is true
PASS /Ҍ/i.test("Ҋ") is false
PASS /҉/i.test("ҋ") is false
PASS /Ҋ/i.test("ҋ") is true
PASS /ҋ/i.test("ҋ") is true
PASS /Ҍ/i.test("ҋ") is false
PASS /[҉-Ҋ]/i.test("ҋ") is true
PASS /[ҋ-Ҍ]/i.test("Ҋ") is true
PASS /ӄ/i.test("Ӆ") is false
PASS /Ӆ/i.test("Ӆ") is true
PASS /ӆ/i.test("Ӆ") is true
PASS /Ӈ/i.test("Ӆ") is false
PASS /ӄ/i.test("ӆ") is false
PASS /Ӆ/i.test("ӆ") is true
PASS /ӆ/i.test("ӆ") is true
PASS /Ӈ/i.test("ӆ") is false
PASS /[ӄ-Ӆ]/i.test("ӆ") is true
PASS /[ӆ-Ӈ]/i.test("Ӆ") is true
PASS successfullyParsed is true

TEST COMPLETE
